# CrawlerAPI/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Crawler API
  version: 1.0.0
  description: |
    OpenAPI specification for the Crawler API. Generated from Spring controller mappings
    under `src/main/java/com/crawlerApi/api`. Many endpoints require an authenticated account
    context (JWT) and return domain-specific DTOs for domains, audits, pages, elements, and test records.
servers:
  - url: http://localhost:9080
    description: Local default server port from application.properties
security:
  - bearerAuth: []
...
paths:
...
  /v1/audits/{id}:
    get:
      tags: [Audits]
      summary: Get audit by id
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200': { description: Audit retrieved }
...
  /v1/domains/audits:
    get:
      tags: [Domains]
      summary: Get most recent domain audit record for host
      description: Get the most recent domain audit record for the given host.
      parameters:
        - in: query
          name: host
          required: true
          schema: { type: string }
          description: Domain host to get the most recent audit for
      responses:
        '200': { description: Domain audit record returned, content: { application/json: { schema: { $ref: '#/components/schemas/DomainAuditRecord' } } } }
        '401': { description: Authentication required }
        '403': { description: Insufficient permissions }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
...
  schemas:
...
    DomainAuditRecord:
      type: object
      description: Domain audit record
      additionalProperties: true
...